Move mempolicy.h to its own module.
After regenerating all the generated files, I noticed that mempolicy.h
was pulling in the errno codes for all architectures, so I think it's
best for it to be in its own module.
