+++ Cargo.toml
@@ -31,6 +31,7 @@ general = []
 if_ether = []
 io_uring = []
 ioctl = []
+mempolicy = []
 net = []
 netlink = []
 prctl = []
