+++ gen/modules/mempolicy.h
@@ -1,0 +1,2 @@
+#include "support.h"
+#include <linux/mempolicy.h>
