+++ src/lib.rs
@@ -227,6 +227,10 @@ pub mod io_uring;
 #[cfg(target_arch = "arm")]
 #[path = "arm/ioctl.rs"]
 pub mod ioctl;
+#[cfg(feature = "mempolicy")]
+#[cfg(target_arch = "arm")]
+#[path = "arm/mempolicy.rs"]
+pub mod mempolicy;
 #[cfg(feature = "net")]
 #[cfg(target_arch = "arm")]
 #[path = "arm/net.rs"]
@@ -263,6 +267,10 @@ pub mod io_uring;
 #[cfg(target_arch = "aarch64")]
 #[path = "aarch64/ioctl.rs"]
 pub mod ioctl;
+#[cfg(feature = "mempolicy")]
+#[cfg(target_arch = "aarch64")]
+#[path = "aarch64/mempolicy.rs"]
+pub mod mempolicy;
 #[cfg(feature = "net")]
 #[cfg(target_arch = "aarch64")]
 #[path = "aarch64/net.rs"]
@@ -299,6 +307,10 @@ pub mod io_uring;
 #[cfg(target_arch = "csky")]
 #[path = "csky/ioctl.rs"]
 pub mod ioctl;
+#[cfg(feature = "mempolicy")]
+#[cfg(target_arch = "csky")]
+#[path = "csky/mempolicy.rs"]
+pub mod mempolicy;
 #[cfg(feature = "net")]
 #[cfg(target_arch = "csky")]
 #[path = "csky/net.rs"]
@@ -335,6 +347,10 @@ pub mod io_uring;
 #[cfg(target_arch = "loongarch64")]
 #[path = "loongarch64/ioctl.rs"]
 pub mod ioctl;
+#[cfg(feature = "mempolicy")]
+#[cfg(target_arch = "loongarch64")]
+#[path = "loongarch64/mempolicy.rs"]
+pub mod mempolicy;
 #[cfg(feature = "net")]
 #[cfg(target_arch = "loongarch64")]
 #[path = "loongarch64/net.rs"]
@@ -371,6 +387,10 @@ pub mod io_uring;
 #[cfg(target_arch = "mips")]
 #[path = "mips/ioctl.rs"]
 pub mod ioctl;
+#[cfg(feature = "mempolicy")]
+#[cfg(target_arch = "mips")]
+#[path = "mips/mempolicy.rs"]
+pub mod mempolicy;
 #[cfg(feature = "net")]
 #[cfg(target_arch = "mips")]
 #[path = "mips/net.rs"]
@@ -407,6 +427,10 @@ pub mod io_uring;
 #[cfg(target_arch = "mips64")]
 #[path = "mips64/ioctl.rs"]
 pub mod ioctl;
+#[cfg(feature = "mempolicy")]
+#[cfg(target_arch = "mips64")]
+#[path = "mips64/mempolicy.rs"]
+pub mod mempolicy;
 #[cfg(feature = "net")]
 #[cfg(target_arch = "mips64")]
 #[path = "mips64/net.rs"]
@@ -443,6 +467,10 @@ pub mod io_uring;
 #[cfg(target_arch = "mips32r6")]
 #[path = "mips32r6/ioctl.rs"]
 pub mod ioctl;
+#[cfg(feature = "mempolicy")]
+#[cfg(target_arch = "mips32r6")]
+#[path = "mips32r6/mempolicy.rs"]
+pub mod mempolicy;
 #[cfg(feature = "net")]
 #[cfg(target_arch = "mips32r6")]
 #[path = "mips32r6/net.rs"]
@@ -479,6 +507,10 @@ pub mod io_uring;
 #[cfg(target_arch = "mips64r6")]
 #[path = "mips64r6/ioctl.rs"]
 pub mod ioctl;
+#[cfg(feature = "mempolicy")]
+#[cfg(target_arch = "mips64r6")]
+#[path = "mips64r6/mempolicy.rs"]
+pub mod mempolicy;
 #[cfg(feature = "net")]
 #[cfg(target_arch = "mips64r6")]
 #[path = "mips64r6/net.rs"]
@@ -515,6 +547,10 @@ pub mod io_uring;
 #[cfg(target_arch = "powerpc")]
 #[path = "powerpc/ioctl.rs"]
 pub mod ioctl;
+#[cfg(feature = "mempolicy")]
+#[cfg(target_arch = "powerpc")]
+#[path = "powerpc/mempolicy.rs"]
+pub mod mempolicy;
 #[cfg(feature = "net")]
 #[cfg(target_arch = "powerpc")]
 #[path = "powerpc/net.rs"]
@@ -551,6 +587,10 @@ pub mod io_uring;
 #[cfg(target_arch = "powerpc64")]
 #[path = "powerpc64/ioctl.rs"]
 pub mod ioctl;
+#[cfg(feature = "mempolicy")]
+#[cfg(target_arch = "powerpc64")]
+#[path = "powerpc64/mempolicy.rs"]
+pub mod mempolicy;
 #[cfg(feature = "net")]
 #[cfg(target_arch = "powerpc64")]
 #[path = "powerpc64/net.rs"]
@@ -587,6 +627,10 @@ pub mod io_uring;
 #[cfg(target_arch = "riscv32")]
 #[path = "riscv32/ioctl.rs"]
 pub mod ioctl;
+#[cfg(feature = "mempolicy")]
+#[cfg(target_arch = "riscv32")]
+#[path = "riscv32/mempolicy.rs"]
+pub mod mempolicy;
 #[cfg(feature = "net")]
 #[cfg(target_arch = "riscv32")]
 #[path = "riscv32/net.rs"]
@@ -623,6 +667,10 @@ pub mod io_uring;
 #[cfg(target_arch = "riscv64")]
 #[path = "riscv64/ioctl.rs"]
 pub mod ioctl;
+#[cfg(feature = "mempolicy")]
+#[cfg(target_arch = "riscv64")]
+#[path = "riscv64/mempolicy.rs"]
+pub mod mempolicy;
 #[cfg(feature = "net")]
 #[cfg(target_arch = "riscv64")]
 #[path = "riscv64/net.rs"]
@@ -659,6 +707,10 @@ pub mod io_uring;
 #[cfg(target_arch = "s390x")]
 #[path = "s390x/ioctl.rs"]
 pub mod ioctl;
+#[cfg(feature = "mempolicy")]
+#[cfg(target_arch = "s390x")]
+#[path = "s390x/mempolicy.rs"]
+pub mod mempolicy;
 #[cfg(feature = "net")]
 #[cfg(target_arch = "s390x")]
 #[path = "s390x/net.rs"]
@@ -695,6 +747,10 @@ pub mod io_uring;
 #[cfg(target_arch = "sparc")]
 #[path = "sparc/ioctl.rs"]
 pub mod ioctl;
+#[cfg(feature = "mempolicy")]
+#[cfg(target_arch = "sparc")]
+#[path = "sparc/mempolicy.rs"]
+pub mod mempolicy;
 #[cfg(feature = "net")]
 #[cfg(target_arch = "sparc")]
 #[path = "sparc/net.rs"]
@@ -731,6 +787,10 @@ pub mod io_uring;
 #[cfg(target_arch = "sparc64")]
 #[path = "sparc64/ioctl.rs"]
 pub mod ioctl;
+#[cfg(feature = "mempolicy")]
+#[cfg(target_arch = "sparc64")]
+#[path = "sparc64/mempolicy.rs"]
+pub mod mempolicy;
 #[cfg(feature = "net")]
 #[cfg(target_arch = "sparc64")]
 #[path = "sparc64/net.rs"]
@@ -767,6 +827,10 @@ pub mod io_uring;
 #[cfg(target_arch = "x86")]
 #[path = "x86/ioctl.rs"]
 pub mod ioctl;
+#[cfg(feature = "mempolicy")]
+#[cfg(target_arch = "x86")]
+#[path = "x86/mempolicy.rs"]
+pub mod mempolicy;
 #[cfg(feature = "net")]
 #[cfg(target_arch = "x86")]
 #[path = "x86/net.rs"]
@@ -803,6 +867,10 @@ pub mod io_uring;
 #[cfg(all(target_arch = "x86_64", target_pointer_width = "64"))]
 #[path = "x86_64/ioctl.rs"]
 pub mod ioctl;
+#[cfg(feature = "mempolicy")]
+#[cfg(all(target_arch = "x86_64", target_pointer_width = "64"))]
+#[path = "x86_64/mempolicy.rs"]
+pub mod mempolicy;
 #[cfg(feature = "net")]
 #[cfg(all(target_arch = "x86_64", target_pointer_width = "64"))]
 #[path = "x86_64/net.rs"]
@@ -839,6 +907,10 @@ pub mod io_uring;
 #[cfg(all(target_arch = "x86_64", target_pointer_width = "32"))]
 #[path = "x32/ioctl.rs"]
 pub mod ioctl;
+#[cfg(feature = "mempolicy")]
+#[cfg(all(target_arch = "x86_64", target_pointer_width = "32"))]
+#[path = "x32/mempolicy.rs"]
+pub mod mempolicy;
 #[cfg(feature = "net")]
 #[cfg(all(target_arch = "x86_64", target_pointer_width = "32"))]
 #[path = "x32/net.rs"]
